pyproject.toml
[tool.poetry]
name = "rpp_sentiment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
Flask-Migrate = "^3.1.0"
Flask-Seeder = "^1.2.0"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
